+++ frontend/lib/api-types.generated.ts
@@ -1748,7 +1748,7 @@ export interface components {
             version: string;
             /**
              * App Version
-             * @default 1.0.13
+             * @default 1.0.14
              */
             app_version: string;
             /** Sandbox Available */
@@ -1776,7 +1776,7 @@ export interface components {
             version: string;
             /**
              * App Version
-             * @default 1.0.13
+             * @default 1.0.14
              */
             app_version: string;
             /** Sandbox Available */
@@ -2162,7 +2162,7 @@ export interface components {
             version: string;
             /**
              * App Version
-             * @default 1.0.13
+             * @default 1.0.14
              */
             app_version: string;
             /**